market_django/requirements.txt
Django==2.0.2
django-configurations==2.0
django-debug-toolbar==1.9.1
django-extensions==2.0.0
Pillow==5.0.0
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
typing==3.6.4
